remove delete for secretary

+++ do-not-delete-this/admin/officials/officials-table.php
@@ -112,9 +112,11 @@ $results = $stmt->fetchAll();
                                 <td><?php echo date('F j, Y', strtotime($resident['date_start'])) ?></td>
                                 <td><?php echo date('F j, Y', strtotime($resident['date_end'])) ?></td>
                                 <td>
-                                    <button type="button" id="deleteBtn" class="focus:outline-none text-white bg-red-700 hover:bg-red-800 focus:ring-4 focus:ring-red-300 font-medium rounded-lg text-sm px-5 py-2.5 mr-2 mb-2 dark:bg-red-600 dark:hover:bg-red-700 dark:focus:ring-red-900">
-                                        <a href="includes/delete-officials.inc.php?id=<?php echo $resident['resident_id'] ?>">Delete</a>
-                                    </button>
+                                    <?php if ($resident['position'] !== 'Barangay Secretary') : ?>
+                                        <button type="button" id="deleteBtn" class="focus:outline-none text-white bg-red-700 hover:bg-red-800 focus:ring-4 focus:ring-red-300 font-medium rounded-lg text-sm px-5 py-2.5 mr-2 mb-2 dark:bg-red-600 dark:hover:bg-red-700 dark:focus:ring-red-900">
+                                            <a href="includes/delete-officials.inc.php?id=<?php echo $resident['resident_id'] ?>">Delete</a>
+                                        </button>
+                                    <?php endif; ?>
                                 </td>
                             </tr>
                         <?php } ?>
